Adds classes Element and Component, remove Store
