+++ src/scripts/components/Element/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Element";
