+++ src/styles/blocks/_button.scss
@@ -14,6 +14,13 @@
   padding: 9.5px 20px;
   min-width: 144px;
   transition: background-color var(--transition-time);
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+
+  img {
+    margin-right: 10px;
+  }
 
   &:hover {
     background-color: darken(#48ba46, 10%);
