+++ webpack/webpack.common.js
@@ -1,32 +1,32 @@
-const Path = require('path');
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+const Path = require("path");
+const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+const CopyWebpackPlugin = require("copy-webpack-plugin");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
 
 module.exports = {
   entry: {
-    app: Path.resolve(__dirname, '../src/scripts/index.js'),
+    app: Path.resolve(__dirname, "../src/scripts/index.js.js"),
   },
   output: {
-    path: Path.join(__dirname, '../build'),
-    filename: 'js/[name].js',
+    path: Path.join(__dirname, "../build"),
+    filename: "js/[name].js",
   },
   optimization: {
     splitChunks: {
-      chunks: 'all',
+      chunks: "all",
       name: false,
     },
   },
   plugins: [
     new CleanWebpackPlugin(),
-    new CopyWebpackPlugin({ patterns: [{ from: Path.resolve(__dirname, '../public'), to: 'public' }] }),
+    new CopyWebpackPlugin({ patterns: [{ from: Path.resolve(__dirname, "../public"), to: "public" }] }),
     new HtmlWebpackPlugin({
-      template: Path.resolve(__dirname, '../src/index.html'),
+      template: Path.resolve(__dirname, "../src/index.js.html"),
     }),
   ],
   resolve: {
     alias: {
-      '~': Path.resolve(__dirname, '../src'),
+      "~": Path.resolve(__dirname, "../src"),
     },
   },
   module: {
@@ -34,14 +34,14 @@ module.exports = {
       {
         test: /\.mjs$/,
         include: /node_modules/,
-        type: 'javascript/auto',
+        type: "javascript/auto",
       },
       {
         test: /\.(ico|jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2)(\?.*)?$/,
         use: {
-          loader: 'file-loader',
+          loader: "file-loader",
           options: {
-            name: '[path][name].[ext]',
+            name: "[path][name].[ext]",
           },
         },
       },
